Moved GEE asset manager
GEE asset manager currently nested in module 1 backend code. Since module 1 will grow, the asset manager is moved to the helper module.
